+++ internal/pkg/dbmigrate/migrations/000092_dispatch_policy_ext.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE dispatch_policies DROP COLUMN unified_template_id;