+++ internal/pkg/dbmigrate/migrations/000095_notify_media_team_id.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE notify_medias DROP INDEX idx_team_id;
+ALTER TABLE notify_medias DROP COLUMN team_id;